feat: analytics isolation
